feat: add the rest of the apps to script-config and update config files
