+++ starter-dev-e2e/cypress/configs/angular-ngrx-scss.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from "cypress";
+
+export default defineConfig({
+  env: {
+    authUrl: "/",
+  },
+  e2e: {
+    baseUrl: "http://localhost:4200",
+  },
+});
