--- starter-dev-e2e/cypress/configs/angular-ngrx-scss.json
@@ -1,8 +1,0 @@
-{
-  "baseUrl": "http://localhost:4200",
-  "startCommand": "yarn start",
-  "integrationFolder": "cypress/integration",
-  "env": {
-    "authUrl": "/"
-  }
-}
