+++ starter-dev-e2e/cypress/configs/cra-rxjs-styled-components.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from "cypress";
+
+export default defineConfig({
+  env: {
+    authUrl: "/redirect",
+  },
+  e2e: {
+    baseUrl: "http://localhost:3000",
+  },
+});
