--- starter-dev-e2e/cypress/configs/cra-rxjs-styled-components.json
@@ -1,8 +1,0 @@
-{
-  "baseUrl": "http://localhost:3000",
-  "startCommand": "npm dev",
-  "integrationFolder": "cypress/integration",
-  "env": {
-    "authUrl": "/redirect"
-  }
-}
