--- starter-dev-e2e/cypress/configs/next-react-query-tailwind.json
@@ -1,11 +1,0 @@
-{
-  "baseUrl": "http://localhost:3000",
-  "startCommand": "yarn dev",
-  "integrationFolder": "cypress/integration",
-  "env": {
-    "authUrl": "/",
-    "JWT_SECRET": "7ZkCMuD69N2WBBDCDzxTTYvsNG8qPJHWeNYe37iZTkmQLUmujZUZ3RnkqP6yfxTU",
-    "JWT_SIGNING_KEY": "{\"kty\":\"oct\",\"kid\":\"6f3fzrvXzvbJ66XfW3QbeU6WfNWm3Z5KwKt5fDSOsdI\",\"alg\":\"HS512\",\"k\":\"mq_ZTza3MtzKDyy7r47-bOeZcvzfsy3VfIsdfT-n5W-KA4mxUy-6Yy9GjISuo3aNfPpzXRDIsM0mS-tFdCuULg\"}",
-    "JWT_ENCRYPTION_KEY": "{\"kty\":\"oct\",\"kid\":\"noSwokthOcoRCtufQ3Eyp_x_cIshHpKBRanEVreZwnc\",\"alg\":\"A256GCM\",\"k\":\"GdCtl3XPLxKfdUQOFyWEPbe27tOPOJ9Z7-RPAdz_Vrc\"}"
-  }
-}
