--- starter-dev-e2e/cypress/configs/remix-gql-tailwind.json
@@ -1,16 +1,0 @@
-{
-  "baseUrl": "http://localhost:3000",
-  "startCommand": "yarn mock",
-  "integrationFolder": "cypress/integration",
-  "env": {
-    "authUrl": "/",
-    "remixCookieOptions": {
-      "name": "__session",
-      "httpOnly": true,
-      "path": "/",
-      "sameSite": "lax",
-      "secrets": ["s3cr3t"],
-      "secure": false
-    }
-  }
-}
