--- starter-dev-e2e/cypress/configs/vite-apollo-vanilla.json
@@ -1,8 +1,0 @@
-{
-  "baseUrl": "http://localhost:3000",
-  "startCommand": "yarn dev",
-  "integrationFolder": "cypress/integration",
-  "env": {
-    "authUrl": "/"
-  }
-}
