+++ starter-dev-e2e/cypress/configs/vue3-ts-quasar.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from "cypress";
+
+export default defineConfig({
+  env: {
+    authUrl: "/redirect",
+  },
+  e2e: {
+    baseUrl: "http://localhost:8080",
+  },
+});
