--- starter-dev-e2e/cypress/configs/vue3-ts-quasar.json
@@ -1,8 +1,0 @@
-{
-  "baseUrl": "http://localhost:8080",
-  "startCommand": "yarn dev",
-  "integrationFolder": "cypress/integration",
-  "env": {
-    "authUrl": "/redirect"
-  }
-}
